+++ tests/conftest.py
@@ -1,0 +1,8 @@
+import tempfile
+from pytest import fixture
+
+
+@fixture
+def tmp_file():
+    with tempfile.NamedTemporaryFile() as out:
+        yield out