+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Merge command: `fiboa merge`
 - Converter for Croatia
 - Converter for Germany, Mecklenburg-Western Pomerania
 - Converter for Germany, Saarland
@@ -21,6 +22,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - `fiboa convert`: Default compression changed from `zstd` to `brotli`
+
+### Fixed
+
+- Datatype conversion from pandas to pyarrow fixed
 
 ## [v0.7.0] - 2024-08-24